Added Quiz Skeleton
Added a quiz skeleton interface. This includes the following.

QuizWindow.fxml
QuizController.java

Used to take the quiz and handle the results.
