+++ bin/application/QuizWindow.fxml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.layout.Pane?>
+<?import javafx.scene.layout.VBox?>
+
+<VBox maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="900.0" prefWidth="750.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.QuizController">
+   <children>
+      <Pane prefHeight="900.0" prefWidth="750.0">
+         <children>
+            <Button fx:id="submitQuizButton" layoutX="307.0" layoutY="793.0" mnemonicParsing="false" onAction="#handleSubmitButtonAction" prefHeight="53.0" prefWidth="136.0" text="Submit" />
+         </children></Pane>
+   </children>
+</VBox>
