+++ src/application/MainGUI.fxml
@@ -7,11 +7,11 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<VBox maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="643.0" prefWidth="927.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.QuizController">
+<VBox maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="643.0" prefWidth="927.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.MainController">
    <children>
       <Pane prefHeight="633.0" prefWidth="925.0">
          <children>
-            <Button fx:id="startQuizButton" layoutX="464.0" layoutY="409.0" mnemonicParsing="false" prefHeight="53.0" prefWidth="132.0" text="Start Quiz" />
+            <Button fx:id="startQuizButton" layoutX="464.0" layoutY="409.0" mnemonicParsing="false" onAction="#handleQuizButtonAction" prefHeight="53.0" prefWidth="132.0" text="Start Quiz" />
             <Button fx:id="houseInfoButton" layoutX="297.0" layoutY="409.0" mnemonicParsing="false" onAction="#handleInfoButtonAction" prefHeight="53.0" prefWidth="144.0" text="House Information" />
             <Label layoutX="25.0" layoutY="14.0" prefHeight="81.0" prefWidth="710.0" text="Welcome to the Harry Potter Housing Quiz">
                <font>
